+++ src/promptForNextVersion.ts
@@ -1,18 +1,51 @@
 import enquirer from 'enquirer'
 import { readPackageUp } from 'read-package-up'
 import semver from 'semver'
+import { RepositoryType } from './getRepositoryType.js'
+import { getNugetVersion } from './nuget.js'
 
-export default async function promptForNextVersion({
+async function getCurrentVersion({
+  type,
   cwd,
 }: {
+  type: RepositoryType
+  cwd: string
+}): Promise<string | undefined> {
+  switch (type.type) {
+    case 'NODE_JS':
+    case 'NPM': {
+      const result = await readPackageUp({
+        cwd,
+      })
+      return result?.packageJson.version
+    }
+    case 'NUGET': {
+      return await getNugetVersion({
+        cwd,
+        relativeProjectFile: type.relativeProjectFile,
+      })
+    }
+  }
+}
+
+export default async function promptForNextVersion({
+  type,
+  cwd,
+}: {
+  type: RepositoryType
   cwd: string
 }): Promise<{
   nextVersion: string
 }> {
-  const result = await readPackageUp({
+  const currentVersion = await getCurrentVersion({
+    type,
     cwd,
   })
-  const currentVersion = semver.valid(result?.packageJson.version)
+  if (!semver.valid(currentVersion)) {
+    throw new Error(
+      `Could not determine current version for repository: ${cwd}`,
+    )
+  }
 
   return await enquirer.prompt<{ nextVersion: string }>({
     type: 'input',